Refactored view functions into its own file
